Update WrapRadianRangeCosSinAtan2.html
Keep a value of Math.PI*2(full circle) as valid and usable by clamping or limiting of wrapping it in a range of -Math.PI to Math.PI. This done by gettng the x and y vector using Math.cos and Math.sin. Then re-getting the angle with Math.atan2(y,x). 
No need for 'complex' unmemorized formula's it seems.